SpaprVio addresses are 32-bit, not 64-bit
spapr-vio addresses are used on POWER platform qemu guests, which are based
on the PAPR specification.  PAPR specifies a number of virtual devices (but
not virtio protocol) which are addressed in an abstract namespace.

Currently, libvirt encodes these addresses as 64-bit values.  This is not
correct: spapr-vio addresses are, and always have been 32-bit.  That's true
both by the PAPR specification and the qemu implementation.

Therefore, change this in libvirt.

This looks like it would be a breaking change, but it actually isn't.
Because these have always been 32-bit at the lower levels, any attempt to
use a value here > 0xffffffff would always have failed in any case, this
will just make it fail earlier and more clearly.

Message-Id: <20190604013840.20291-1-david@gibson.dropbear.id.au>

+++ src/conf/device_conf.h
@@ -101,7 +101,7 @@ struct _virDomainDeviceUSBAddress {
 typedef struct _virDomainDeviceSpaprVioAddress virDomainDeviceSpaprVioAddress;
 typedef virDomainDeviceSpaprVioAddress *virDomainDeviceSpaprVioAddressPtr;
 struct _virDomainDeviceSpaprVioAddress {
-    unsigned long long reg;
+    unsigned long reg;
     bool has_reg;
 };
